requirements.txt
mypy==0.961
dnspython==2.2.1
ldap3==2.9.1
PyQt5==5.15.7
pycryptodome==3.15.0
clipboard==0.0.4
